# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Helpdesk API - REST",
        "description": "API REST del Sistema Helpdesk. Migración de GraphQL a REST con autenticación JWT.",
        "contact": {
            "name": "API Support",
            "email": "support@helpdesk.local"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Development Server"
        }
    ],
    "paths": {
...
        "/api/company-requests": {
            "get": {
                "tags": [
                    "Company Requests"
                ],
                "summary": "Listar solicitudes de empresas",
                "description": "Retorna lista paginada de solicitudes de empresas. Requiere rol PLATFORM_ADMIN. Incluye eager loading de reviewer y createdCompany.",
                "operationId": "list_company_requests",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filtrar por estado de la solicitud",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "PENDING",
                                "APPROVED",
                                "REJECTED"
                            ]
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Buscar por nombre de empresa",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "Campo para ordenar",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "order",
                        "in": "query",
                        "description": "Dirección de ordenamiento",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "asc",
                                "desc"
                            ]
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Número de página",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "minimum": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de solicitudes de empresas",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "string",
                                                        "format": "uuid"
                                                    },
                                                    "requestCode": {
                                                        "type": "string"
                                                    },
                                                    "companyName": {
                                                        "type": "string"
                                                    },
                                                    "legalName": {
                                                        "type": "string",
                                                        "nullable": true
                                                    },
                                                    "adminEmail": {
                                                        "type": "string",
                                                        "format": "email"
                                                    },
                                                    "status": {
                                                        "type": "string",
                                                        "enum": [
                                                            "PENDING",
                                                            "APPROVED",
                                                            "REJECTED"
                                                        ]
                                                    },
                                                    "createdAt": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "meta": {
                                            "properties": {
                                                "total": {
                                                    "type": "integer"
                                                },
                                                "current_page": {
                                                    "type": "integer"
                                                },
                                                "last_page": {
                                                    "type": "integer"
                                                },
                                                "per_page": {
                                                    "type": "integer"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "links": {
                                            "properties": {
                                                "first": {
                                                    "type": "string",
                                                    "nullable": true
                                                },
                                                "last": {
                                                    "type": "string",
                                                    "nullable": true
                                                },
                                                "prev": {
                                                    "type": "string",
                                                    "nullable": true
                                                },
                                                "next": {
                                                    "type": "string",
                                                    "nullable": true
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Sin autenticación"
                    },
                    "403": {
                        "description": "Sin permisos - requiere rol PLATFORM_ADMIN"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Company Requests"
                ],
                "summary": "Crear solicitud de empresa",
                "description": "Endpoint público para crear solicitud de nueva empresa. Rate limit: 3 solicitudes por hora. Validación automática de email duplicado en solicitudes pendientes.",
                "operationId": "create_company_request",
                "requestBody": {
                    "description": "Datos de la solicitud de empresa",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "company_name",
                                    "admin_email",
                                    "business_description",
                                    "industry_type"
                                ],
                                "properties": {
                                    "company_name": {
                                        "description": "Nombre de la empresa",
                                        "type": "string",
                                        "maxLength": 200,
                                        "minLength": 2
                                    },
                                    "legal_name": {
                                        "description": "Nombre legal de la empresa",
                                        "type": "string",
                                        "maxLength": 200,
                                        "minLength": 2,
                                        "nullable": true
                                    },
                                    "admin_email": {
                                        "description": "Email del administrador",
                                        "type": "string",
                                        "format": "email",
                                        "maxLength": 255
                                    },
                                    "business_description": {
                                        "description": "Descripción del negocio",
                                        "type": "string",
                                        "maxLength": 2000,
                                        "minLength": 50
                                    },
                                    "website": {
                                        "description": "Sitio web de la empresa",
                                        "type": "string",
                                        "format": "uri",
                                        "maxLength": 255,
                                        "nullable": true
                                    },
                                    "industry_type": {
                                        "description": "Tipo de industria",
                                        "type": "string",
                                        "maxLength": 100
                                    },
                                    "estimated_users": {
                                        "description": "Número estimado de usuarios",
                                        "type": "integer",
                                        "maximum": 10000,
                                        "minimum": 1,
                                        "nullable": true
                                    },
                                    "contact_address": {
                                        "description": "Dirección de contacto",
                                        "type": "string",
                                        "maxLength": 255,
                                        "nullable": true
                                    },
                                    "contact_city": {
                                        "description": "Ciudad",
                                        "type": "string",
                                        "maxLength": 100,
                                        "nullable": true
                                    },
                                    "contact_country": {
                                        "description": "País",
                                        "type": "string",
                                        "maxLength": 100,
                                        "nullable": true
                                    },
                                    "contact_postal_code": {
                                        "description": "Código postal",
                                        "type": "string",
                                        "maxLength": 20,
                                        "nullable": true
                                    },
                                    "tax_id": {
                                        "description": "Identificador fiscal",
                                        "type": "string",
                                        "maxLength": 50,
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Solicitud de empresa creada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "string",
                                            "format": "uuid"
                                        },
                                        "requestCode": {
                                            "type": "string"
                                        },
                                        "companyName": {
                                            "type": "string"
                                        },
                                        "adminEmail": {
                                            "type": "string",
                                            "format": "email"
                                        },
                                        "status": {
                                            "type": "string",
                                            "enum": [
                                                "PENDING"
                                            ]
                                        },
                                        "createdAt": {
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación - datos inválidos o email duplicado"
                    },
                    "429": {
                        "description": "Rate limit excedido - máximo 3 solicitudes por hora"
                    }
                }
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "JWT Bearer Token. Include in Authorization header: Bearer <token>",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}